refactor: introduce AgentModel and AgentContext for model/view separation
Extracts all agent orchestration logic out of the React component tree
into a plain TypeScript class (AgentModel) and a React ViewModel layer
(AgentContext), following the MVC pattern described in the bandarrame_rs
admin-agent spec.

- src/lib/agent/types.ts: StreamItem/ChildItem type definitions moved
  here so AgentModel can use them without importing from a React component
- src/lib/agent/AgentModel.ts: plain TS class (no React imports); owns
  the Conversation, full streaming loop, conversation history init, and
  exposes items/isLoading/sendMessage/cancel/subscribe
- src/context/AgentContext.tsx: React ViewModel; creates EditorTools,
  WorkspaceTools, factory, and registry (all moved out of App.tsx);
  instantiates AgentModel via useMemo, subscribes to state changes,
  and exposes sendMessage/cancel/items/isLoading to the tree
- App.tsx: agent setup entirely removed; layouts take no props; wrapped
  in AgentContextProvider
- ChatSidebar.tsx: conversation prop removed; reads from useAgentContext();
  handleSend is now a thin dispatcher; expandedThoughts managed via
  useEffect watching items
- ChatSidebar.test.tsx: updated to mock useAgentContext and test
  display-level behaviour via static items
